Clarify empty window messaging when scheduler runs mid-window
